Generate Java proxy class fields as member accessors
Hacky, proof-of-concept code.

This replaces the ugly JavaBean type getters/setters with generic fields that
expose a get()/set() method pair. This pair is 'wired' to the method body code
that would have been generated to support the JavaBean getters/setters.

Essentially, having a field to represent the structure member variable makes the
resulting Java code a bit more attractive to the eye, as it is easier to see the
correspondence between the native struct member and the proxy class field; they
share the same name.