Generate 'address-of' struct member accessor
Expose an additional accessor for getting the address of a struct member.

Hacky, proof-of-concept code. Only tested with Java for wrapping C structs and
almost certainly breaks other language modules. Not tested with C++ classes,
directors, smart pointers etc.

With some extra effort, this can controlled via a feature. Maybe language
modules supporting that feature should report it so that the related code in
lang.cxx can execute conditionally (otherwise enabling the future can lead to
unexpected results for language modules that do not support it).